Replace `ThunkExpression#simplify()` with visitor
Functionality to simplify an expression does not belong in thunx-model
classes, which should be almost purely data-containers without behavior.

The existing visitor pattern can be used to implement
`ThunkReducerVisitor` which will reduce the expression as much as
possible.
