+++ thunx-model/src/main/java/eu/xenit/contentcloud/thunx/predicates/model/ThunkExpressionVisitor.java
@@ -6,4 +6,5 @@ public interface ThunkExpressionVisitor<T> {
     T visit(FunctionExpression<?> functionExpression);
     T visit(SymbolicReference symbolicReference);
     T visit(Variable variable);
+
 }
