+++ thunx-model/src/test/java/eu/xenit/contentcloud/thunx/predicates/model/LogicalOperationTest.java
@@ -20,7 +20,6 @@ class LogicalOperationTest {
         var disjunction = LogicalOperation.disjunction(rule1, rule2, rule3);
 
         assertThat(disjunction.getOperator()).isEqualTo(Operator.OR);
-        assertThat(Scalar.maybeValue(disjunction.simplify())).isNotPresent();
         Assertions.assertThat(disjunction.getTerms()).containsExactly(rule1, rule2, rule3);
         assertThat(disjunction).isEqualTo(LogicalOperation.disjunction(rule1, rule2, rule3));
     }
@@ -33,99 +32,11 @@ class LogicalOperationTest {
         var conjunction = LogicalOperation.conjunction(rule1, rule2);
 
         assertThat(conjunction.getOperator()).isEqualTo(Operator.AND);
-        assertThat(Scalar.maybeValue(conjunction.simplify())).isNotPresent();
         Assertions.assertThat(conjunction.getTerms()).hasSize(2);
 
         assertThat(conjunction).isEqualTo(LogicalOperation.conjunction(rule1, rule2));
     }
 
-    @Test
-    void resolvableDisjunction() {
-        var disjunction = LogicalOperation.disjunction(Scalar.of(true));
-
-        var simplified = disjunction.simplify();
-        var resolved = Scalar.maybeValue(simplified);
-        assertThat(resolved).isPresent().contains(true);
-    }
-
-    @Test
-    void resolvableDisjunctions_someResolvableTrue() {
-        var disjunction = LogicalOperation.disjunction(
-                Scalar.of(true),
-                Comparison.areEqual(SymbolicReference.of("document.security"), Scalar.of(5))
-        );
-
-        var simplified = disjunction.simplify();
-        var resolved = Scalar.maybeValue(simplified);
-        assertThat(resolved).isPresent().contains(true);
-    }
-
-    @Test
-    void resolvableDisjunctions_someResolvableFalse() {
-        var disjunction = LogicalOperation.disjunction(
-                Scalar.of(false),
-                Comparison.areEqual(SymbolicReference.of("document.security"), Scalar.of(5))
-        );
-
-        var simplified = disjunction.simplify();
-        assertThat(simplified).hasToString("EQ(document.security, 5)");
-
-        var resolved = Scalar.maybeValue(simplified);
-        assertThat(resolved).isNotPresent();
-    }
-
-    @Test
-    void resolvableDisjunctions_empty() {
-        var disjunction = LogicalOperation.disjunction(Collections.emptyList());
-
-        var simplified = disjunction.simplify();
-        var resolved = Scalar.maybeValue(simplified);
-        assertThat(resolved).isPresent().contains(false);
-    }
-
-    @Test
-    void resolvableConjunction() {
-        var conjunction = LogicalOperation.conjunction(Scalar.of(true));
-
-        var simplified = conjunction.simplify();
-        var resolved = Scalar.maybeValue(simplified);
-        assertThat(resolved).isPresent().contains(true);
-    }
-
-    @Test
-    void resolvableConjunctions_someResolvableFalse() {
-        var conjunction = LogicalOperation.conjunction(
-                Scalar.of(false),
-                Comparison.areEqual(SymbolicReference.of("document.security"), Scalar.of(5))
-        );
-
-        var simplified = conjunction.simplify();
-        var resolved = Scalar.maybeValue(simplified);
-        assertThat(resolved).isPresent().contains(false);
-    }
-
-    @Test
-    void resolvableConjunctions_someResolvableTrue() {
-        var conjunction = LogicalOperation.conjunction(
-                Scalar.of(true),
-                Comparison.areEqual(SymbolicReference.of("document.security"), Scalar.of(5))
-        );
-
-        var simplified = conjunction.simplify();
-        assertThat(simplified).hasToString("EQ(document.security, 5)");
-
-        var resolved = Scalar.maybeValue(simplified);
-        assertThat(resolved).isNotPresent();
-    }
-
-    @Test
-    void resolvableConjunctions_empty() {
-        var conjunction = LogicalOperation.conjunction(Collections.emptyList());
-
-        var simplified = conjunction.simplify();
-        var resolved = Scalar.maybeValue(simplified);
-        assertThat(resolved).isPresent().contains(true);
-    }
 
     @Test
     void conjunction_toString() {
