+++ thunx-visitor-reducer/gradle.properties
@@ -1,0 +1,1 @@
+description=Library to reduce thunx expressions to their minimal representation
