+++ thunx-visitor-reducer/src/test/java/eu/xenit/contentcloud/thunx/visitor/reducer/ComparisonFunctionReducerTest.java
@@ -1,0 +1,7 @@
+package eu.xenit.contentcloud.thunx.visitor.reducer;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class ComparisonFunctionReducerTest {
+
+}
